safari-600.7-branch/LayoutTests/fast/canvas/canvas-2d-imageData-create-nonfinite-expected.txt

Test the argument bounds of canvas createImageData.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.createImageData(Infinity, Infinity) threw exception TypeError: Type error.
PASS ctx.createImageData(Infinity, 10) threw exception TypeError: Type error.
PASS ctx.createImageData(-Infinity, 10) threw exception TypeError: Type error.
PASS ctx.createImageData(10, Infinity) threw exception TypeError: Type error.
PASS ctx.createImageData(10, -Infinity) threw exception TypeError: Type error.
PASS ctx.createImageData(NaN, 10) threw exception TypeError: Type error.
PASS ctx.createImageData(10, NaN) threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
